Launch toggle intents in a floating window (1/2)
If option is enabled, a toggles intent will launch in a floating window. This mostly affects
toggle long click actions as short click actions usually enable or disable or perform an action.

Im sure theres a faster, cleaner way to do this but I am a noob at this. Until then, this works.

Change-Id: Ia98b787252ac0bc9c4fb3a1dab1213032917dd1d